+++ src/fastlife/config/configurator.py
@@ -33,14 +33,14 @@ from fastlife.middlewares.base import AbstractMiddleware
 from fastlife.routing.route import Route
 from fastlife.routing.router import Router
 from fastlife.security.csrf import check_csrf
-from fastlife.services.policy import check_permission
 from fastlife.services.registry import DefaultRegistry, TRegistry
+from fastlife.services.security_policy import check_permission
 from fastlife.settings import Settings
 from fastlife.shared_utils.resolver import resolve, resolve_maybe_relative
 from fastlife.templates.inline import is_inline_template_returned
 
 if TYPE_CHECKING:
-    from fastlife.security.policy import AbstractSecurityPolicy  # coverage: ignore
+    from fastlife.services.security_policy import AbstractSecurityPolicy
     from fastlife.services.templates import (
         AbstractTemplateRendererFactory,  # coverage: ignore
     )
@@ -518,7 +518,7 @@ class GenericConfigurator(Generic[TRegistry]):
             parameters.
         :param endpoint: the function that will reveive the request.
         :param permission: a permission to validate by the
-            {class}`Security Policy <fastlife.security.policy.AbstractSecurityPolicy>`.
+            {class}`Security Policy <fastlife.services.security_policy.AbstractSecurityPolicy>`.
         :param status_code: customize response status code.
         :param methods: restrict route to a list of http methods.
         :return: the configurator.