+++ src/fastlife/config/views.py
@@ -47,7 +47,7 @@ def view_config(
     :param template: the template rendered by the
         {class}`fastlife.services.templates.AbstractTemplateRenderer`.
     :param permission: a permission to validate by the
-        {class}`Security Policy <fastlife.security.policy.AbstractSecurityPolicy>`.
+        {class}`Security Policy <fastlife.services.security_policy.AbstractSecurityPolicy>`.
     :param status_code: customize response status code.
     :param methods: restrict route to a list of http methods.