+++ tests/fastlife_app/views/app/admin/secured.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from fastapi import Depends, Query, Response
 
 from fastlife import FormModel, JinjaXTemplate, Request, form_model, view_config
-from fastlife.security.policy import Forbidden
+from fastlife.services.security_policy import Forbidden
 from tests.fastlife_app.service.uow import AuthenticatedUser
 from tests.fastlife_app.views.app.home import Person